Move tray stuff into tray.js
