+++ package.json
@@ -23,7 +23,6 @@
         "chokidar": "^5.0.0",
         "compare-versions": "^6.1.1",
         "electron-conf": "^1.3.0",
-        "electron-prompt": "^1.7.0",
         "electron-store": "^11.0.2",
         "feather-icons": "^4.29.2",
         "feed": "^4.2.2",
