+++ src/app/index.js
@@ -1,56 +1,30 @@
-import * as fs from "node:fs"
 import * as path from "node:path"
 import { exec } from "node:child_process"
 import { platform } from "node:os"
 
-import _ from "lodash"
-import prompt from "electron-prompt"
 import winston from "winston"
 import { BugSplatNode as BugSplat } from "bugsplat-node"
-import tiny from "tiny-json-http"
 
 import {
-    APP_PATH,
-    USER_DATA_PATH,
     APP_SETTINGS,
     openExternalUrl,
     showMessageBox,
-    showNotification,
-    showPrompt,
-    showFilePicker,
-    createTray,
     LOG_PATH,
 } from "./electron.js"
 import { renderFormInWindow, openPageInWindow } from "./window.js"
 import config from "../config/index.js"
 import projects from "./projects.js"
-import { deploy, presets, IS_PLUS_MODE, getNeocitiesApiKey } from "../deploy.js"
+import { deploy, getNeocitiesApiKey } from "../deploy.js"
 import strings from "../config/strings.js"
 import urls from "../config/urls.js"
 import { build } from "../site-generator.js"
-import {
-    app,
-    globalShortcut,
-    crashReporter,
-    ipcMain,
-    clipboard,
-    dialog,
-    shell,
-    Menu,
-} from "electron"
+import { app, globalShortcut, crashReporter, ipcMain, Menu } from "electron"
 import { resolveHandle as resolveBlueskyHandle } from "../bluesky/main.js"
-import {
-    createNewProject,
-    activeProject,
-    setActiveProject,
-    getProjectStarters,
-} from "../index.js"
-import { checkVersion, CURRENT_VERSION, versionIsCurrent } from "./version.js"
+import { createNewProject, activeProject } from "../index.js"
+import { checkVersion, CURRENT_VERSION } from "./version.js"
+import { clearTrayProject, initializeTray, updateTrayTitle } from "./tray.js"
 
 let bugsplat = null
-
-let tray = null
-let trayMenu = null
 
 configureCrashReporting()
 
@@ -73,17 +47,6 @@ app.whenReady().then(() => {
 
     projects.cleanup()
 
-    tray = createTray()
-    tray.on("click", (event) => {
-        // TODO rework tray updates so that we don't to do this on click, because it doesn't work on windows right-click
-        updateTrayMenu(event.shiftKey || config.DEV_MODE)
-        // allow opening context menu with left click on windows
-        if (platform() === "win32") {
-            tray.popUpContextMenu()
-        }
-    })
-    tray.setIgnoreDoubleClickEvents(true)
-
     globalShortcut.register("CommandOrControl+Alt+R", clearConfig)
 
     // keeps app open with no real windows
@@ -93,8 +56,7 @@ app.whenReady().then(() => {
         openExternalUrl(urls.tutorial)
     }
 
-    updateTrayTitle()
-    updateTrayMenu()
+    initializeTray()
 
     checkVersion()
 
@@ -113,219 +75,6 @@ app.on("web-contents-created", (event, contents) => {
         return { action: "deny" }
     })
 })
-
-function updateTrayMenu(isDebugMode = config.DEV_MODE) {
-    const isProjectLoaded = !!activeProject
-    const loadedProjectMeta = isProjectLoaded ? activeProject.config : {}
-    const loadedProjectRoot = isProjectLoaded ? activeProject.paths.ROOT : ""
-
-    const DEPLOY_META = isProjectLoaded && activeProject.secrets?.deployment
-    const BSKY_META =
-        isProjectLoaded && activeProject.secrets?.integrations?.bluesky
-    const bskyAutoPostEnabled = APP_SETTINGS.get("settings.bskyAutoPost")
-
-    trayMenu = Menu.buildFromTemplate([
-        {
-            label: strings.app.titleWithVersion(CURRENT_VERSION),
-            enabled: false,
-        },
-        {
-            label: strings.menu.updateAvailable,
-            visible: !versionIsCurrent(),
-            click: () => openExternalUrl(urls.itch),
-        },
-        { type: "separator" },
-        {
-            id: "title",
-            label: isProjectLoaded
-                ? activeProject.title
-                : strings.projects.notLoaded,
-            type: "submenu",
-            submenu: getProjectsSubmenu(),
-        },
-        {
-            label: strings.menu.openPreview,
-            enabled: isProjectLoaded,
-            click: () => openExternalUrl(urls.localPreview),
-        },
-        { type: "separator" },
-        {
-            label: strings.menu.openEditor,
-            enabled: isProjectLoaded,
-            click: () => {
-                logger.info(strings.logMsg.tryEditor(config.EDITOR_COMMAND))
-
-                exec(
-                    `${config.EDITOR_COMMAND} "${loadedProjectRoot}"`,
-                    (error, stdout, stderr) => {
-                        if (error) {
-                            logger.error(error)
-                            showMessageBox(strings.popups.codiumError)
-                        }
-                        if (stdout) {
-                            logger.info(stdout)
-                        }
-                        if (stderr) {
-                            logger.error(stderr)
-                        }
-                    },
-                )
-            },
-        },
-        {
-            label: strings.menu.openFolder,
-            enabled: isProjectLoaded,
-            click: () => shell.openPath(loadedProjectRoot),
-        },
-        { type: "separator" },
-        {
-            id: "deploy",
-            label: strings.menu.deploy(DEPLOY_META?.provider),
-            visible: !!DEPLOY_META && Object.keys(presets).length > 0,
-            click: () => {
-                // TODO project-level setting to turn off confirmation prompt?
-                const CLICKED_ID = showPrompt(
-                    strings.popups.confirmDeployment.message(
-                        activeProject.title,
-                        DEPLOY_META.provider,
-                    ),
-                    "warning",
-                )
-
-                if (CLICKED_ID == 0) {
-                    // TODO prompt for password if necessary
-                    deploy()
-                } else {
-                    logger.info(strings.deployment.finish.cancel)
-                }
-            },
-        },
-        {
-            label: strings.menu.configDeployment,
-            type: "submenu",
-            enabled: Object.keys(presets).length > 0 && isProjectLoaded,
-            visible: !DEPLOY_META,
-            submenu: Menu.buildFromTemplate(
-                Object.keys(presets).map((key) => {
-                    return {
-                        label: key,
-                        click: () => renderFormInWindow(key),
-                    }
-                }),
-            ),
-        },
-        {
-            label: bskyAutoPostEnabled
-                ? strings.menu.bskyAutoPost.enabled(BSKY_META?.handle)
-                : strings.menu.bskyAutoPost.disabled,
-            visible: !!BSKY_META && IS_PLUS_MODE,
-            click: () =>
-                APP_SETTINGS.set(
-                    "settings.bskyAutoPost",
-                    !APP_SETTINGS.get("settings.bskyAutoPost"),
-                ),
-        },
-        {
-            label: strings.menu.configBsky,
-            enabled: IS_PLUS_MODE && isProjectLoaded,
-            visible: !BSKY_META,
-            click: () => showHtmlForm("bluesky"),
-        },
-        {
-            label: strings.menu.upgrade,
-            visible: Object.keys(presets).length == 0,
-            click: () => openExternalUrl(urls.itch),
-        },
-        { type: "separator" },
-        {
-            label: strings.menu.settings.title,
-            type: "submenu",
-            submenu: getSettingsMenu(),
-        },
-        {
-            label: strings.menu.support.title,
-            type: "submenu",
-            submenu: Menu.buildFromTemplate([
-                {
-                    label: strings.menu.support.checkForUpdates,
-                    click: () => checkVersion(),
-                },
-                {
-                    label: strings.menu.support.openDiscord,
-                    click: () => openExternalUrl(urls.discord),
-                },
-                {
-                    label: strings.menu.support.sendEmail,
-                    click: () => {
-                        if (bugsplat) {
-                            bugsplat.post(new Error("user prompted email"))
-                        }
-                        openExternalUrl(urls.supportMailto)
-                    },
-                },
-            ]),
-        },
-        {
-            label: strings.menu.debug.title,
-            visible: isDebugMode,
-            type: "submenu",
-            submenu: Menu.buildFromTemplate([
-                {
-                    label: strings.menu.debug.copyLog,
-                    click: () => {
-                        const LOG_CONTENTS = fs.readFileSync(LOG_PATH, "utf-8")
-                        clipboard.writeText(LOG_CONTENTS)
-                    },
-                },
-                {
-                    label: strings.menu.debug.openUserData,
-                    click: () => shell.openPath(USER_DATA_PATH),
-                },
-                {
-                    label: strings.menu.debug.deleteSecrets,
-                    click: () => {
-                        const CLICKED_ID = showPrompt(
-                            strings.popups.confirmDeleteSecrets.message,
-                            "warning",
-                            [
-                                strings.popups.confirmDeleteSecrets.confirm,
-                                strings.popups.confirmDeleteSecrets.cancel,
-                            ],
-                        )
-                        if (CLICKED_ID == 0) {
-                            fs.rmSync(activeProject.paths.SECRETS_FILE)
-                        }
-                    },
-                },
-                {
-                    label: strings.menu.debug.clearConfig,
-                    click: clearConfig,
-                },
-            ]),
-        },
-        {
-            label: strings.menu.exit,
-            role: "quit",
-        },
-    ])
-
-    tray.setContextMenu(trayMenu)
-}
-
-function updateTrayTitle() {
-    let displayTitle = strings.projects.notLoaded
-
-    if (activeProject) {
-        displayTitle = activeProject.title
-    }
-
-    tray.setToolTip(displayTitle)
-    tray.setTitle(
-        APP_SETTINGS.get("settings.showProjectTitleInMenubar")
-            ? displayTitle
-            : "",
-    )
-}
 
 function configureCrashReporting() {
     const javaScriptErrorHandler = async (error) => {
@@ -360,157 +109,13 @@ function configureCrashReporting() {
     }
 }
 
-function clearConfig() {
+export function clearConfig() {
     logger.info(strings.logMsg.configClearTry)
     APP_SETTINGS.clear()
     projects.activeIndex = -1
-    tray.setToolTip(strings.projects.notLoaded)
-    tray.setTitle(strings.projects.notLoaded)
+    clearTrayProject()
     showMessageBox(strings.app.configClear)
     logger.info(strings.logMsg.configClearSuccess)
-}
-
-function getSettingsMenu() {
-    const callbacks = {
-        showProjectTitleInMenubar: updateTrayTitle,
-        submitCrashLogs: () => {
-            if (APP_SETTINGS.get("settings.submitCrashLogs")) {
-                let clickedId = showPrompt(
-                    strings.popups.disableCrashReporting.message,
-                    "warning",
-                    [
-                        strings.popups.disableCrashReporting.confirm,
-                        strings.popups.disableCrashReporting.cancel,
-                    ],
-                )
-
-                if (clickedId == 0) {
-                    APP_SETTINGS.set("settings.submitCrashLogs", false)
-                    configureCrashReporting()
-                }
-            } else {
-                APP_SETTINGS.set("settings.submitCrashLogs", true)
-                configureCrashReporting()
-            }
-        },
-    }
-
-    const requireConfirmation = ["submitCrashLogs"]
-    const hidden = ["showAssistant", "bskyAutoPost"]
-
-    let items = _.map(APP_SETTINGS.defaultValues.settings, (v, k) => {
-        if (hidden.includes(k)) {
-            return
-        }
-
-        return {
-            label: strings.menu.settings[k],
-            type: "checkbox",
-            checked: APP_SETTINGS.get(`settings.${k}`),
-            click: () => {
-                if (!requireConfirmation.includes(k)) {
-                    APP_SETTINGS.set(
-                        `settings.${k}`,
-                        !APP_SETTINGS.get(`settings.${k}`),
-                    )
-                }
-
-                if (k in callbacks) {
-                    callbacks[k]()
-                }
-            },
-        }
-    })
-
-    items = _.without(items, undefined)
-
-    return Menu.buildFromTemplate(items)
-}
-
-function getProjectsSubmenu() {
-    const PROJECT_MENU_ITEM = (projectPath, index) => {
-        const PROJECT = projects.getFromPath(projectPath)
-        const RELATIVE_PATH = path.relative(APP_PATH, PROJECT.paths.ROOT)
-        const IS_STARTER =
-            !RELATIVE_PATH.startsWith("..") && !path.isAbsolute(RELATIVE_PATH)
-        const PROJECT_TITLE = `${IS_STARTER ? "📝 " : ""}${PROJECT.title}`
-
-        return {
-            label: PROJECT_TITLE,
-            type: "radio",
-            checked: index == projects.activeIndex,
-            click: () => {
-                projects.activeIndex = index
-
-                // TODO dedupe
-                let displayTitle = APP_SETTINGS.get(
-                    "settings.showProjectTitleInMenubar",
-                )
-                    ? PROJECT_TITLE
-                    : ""
-                tray.setToolTip(displayTitle)
-                tray.setTitle(displayTitle)
-            },
-        }
-    }
-
-    let menuTemplate = [
-        ...projects.list.map(PROJECT_MENU_ITEM),
-        { type: "separator" },
-        {
-            label: strings.menu.projects.create,
-            click: async function () {
-                const browserWindow = await openPageInWindow("new-project")
-                // Send list of starters to form
-                browserWindow.webContents.send(
-                    "starters-list",
-                    Object.keys(getProjectStarters()),
-                )
-                ipcMain.handle("pick-directory", () =>
-                    handlePickDirectory(browserWindow),
-                )
-                browserWindow.on("closed", () =>
-                    ipcMain.removeHandler("pick-directory"),
-                )
-            },
-        },
-        {
-            label: strings.menu.projects.import,
-            click: () => {
-                let pickedPaths = showFilePicker({
-                    filters: [
-                        { name: strings.app.projectFile, extensions: ["yaml"] },
-                    ],
-                    properties: ["openFile"],
-                })
-
-                if (!pickedPaths) {
-                    return
-                }
-
-                projects.add(path.dirname(pickedPaths[0]))
-                projects.activeIndex = projects.list.length - 1
-            },
-        },
-    ]
-
-    // if (isDev()) {
-    //     menuTemplate.unshift(PROJECT_STARTERS.map(PROJECT_MENU_ITEM)) // no index?
-    // }
-
-    return Menu.buildFromTemplate(menuTemplate)
-}
-
-async function handlePickDirectory(attachToWindow) {
-    const { canceled, filePaths } = await dialog.showOpenDialog(
-        attachToWindow,
-        {
-            properties: ["openDirectory"],
-        },
-    )
-    if (!canceled) {
-        return filePaths[0]
-    }
 }
 
 ipcMain.on("form", async function (event, formData) {
