+++ src/app/version.js
@@ -1,11 +1,11 @@
 import { app } from "electron"
+import tiny from "tiny-json-http"
 import { compareVersions } from "compare-versions"
 import { showNotification } from "./electron.js"
 import config from "../config/index.js"
 import strings from "../config/strings.js"
 
 import pkg from "../../package.json" with { type: "json" }
-
 
 // TODO clean up version/update stuff
